Add config params support for create and update command
