+++ packages/zcli-apps/src/commands/apps/server.ts
@@ -6,6 +6,7 @@ import * as cors from 'cors'
 import { buildAppJSON } from '../../lib/buildAppJSON'
 import { Installation } from '../../types'
 import { getAppPaths } from '../../utils/shared'
+import { DEFAULT_APPS_CONFIG_FILE } from '../../constants'
 
 const logMiddleware = morgan((tokens, req, res) =>
   `${chalk.green(tokens.method(req, res))} ${tokens.url(req, res)} ${chalk.bold(tokens.status(req, res))}`
@@ -38,7 +39,7 @@ export default class Server extends Command {
   async run () {
     const { flags } = this.parse(Server)
     const port = parseInt(flags.port)
-    const config = 'zcli.apps.config.json'
+    const config = DEFAULT_APPS_CONFIG_FILE
     const { logs: tailLogs, bind: host } = flags
     const { argv: appDirectories } = this.parse(Server)
