+++ packages/zcli-apps/src/commands/apps/update.ts
@@ -8,8 +8,9 @@ import { getUploadJobStatus } from '../../utils/uploadApp'
 import { promptAndGetSettings, uploadAppPkg, deployApp } from '../../utils/createApp'
 import { getManifestFile } from '../../utils/manifest'
 import { createAppPkg } from '../../lib/package'
-import { Manifest, Installations } from '../../types'
+import { Manifest, Installations, ZcliConfigFileContent } from '../../types'
 import { validateAppPath } from '../../lib/appPath'
+import { getAppSettings } from '../../utils/getAppSettings'
 
 export default class Update extends Command {
   static description = 'updates an existing private app in the Zendesk products specified in the apps manifest file.'
@@ -21,22 +22,24 @@ export default class Update extends Command {
   static strict = false
 
   getAppID (appPath: string) {
-    const allConfigs = getAllConfigs(appPath, 'zcli.apps.config.json')
+    const allConfigs = getAllConfigs(appPath)
     const app_id = allConfigs ? allConfigs.app_id : undefined
     if (!app_id) { throw new CLIError(chalk.red('App ID not found')) }
     return app_id
   }
 
-  async installApp (appId: string, uploadId: number, appPath: string, manifest: Manifest) {
+  async installApp (appConfig: ZcliConfigFileContent, uploadId: number, appPath: string, manifest: Manifest) {
     cli.action.start('Deploying app')
-    const { job_id } = await deployApp('PUT', `api/v2/apps/${appId}`, uploadId)
+    const { job_id } = await deployApp('PUT', `api/v2/apps/${appConfig.app_id}`, uploadId)
 
     try {
       const { app_id }: any = await getUploadJobStatus(job_id, appPath)
       cli.action.stop('Deployed')
 
       const installations: Installations = await request.requestAPI('/api/v2/apps/installations.json', {}, true)
-      const settings = manifest.parameters ? await promptAndGetSettings(manifest.parameters) : {}
+
+      const configParams = appConfig?.parameters || {} // if there are no parameters in the config, just attach an empty object
+      const settings = manifest.parameters ? await getAppSettings(manifest, configParams) : {}
       const installation_id = installations.installations.filter(i => i.app_id === app_id)[0].id
       const updated = await request.requestAPI(`/api/v2/apps/installations/${installation_id}.json`, {
         method: 'PUT',
@@ -64,8 +67,7 @@ export default class Update extends Command {
       validateAppPath(appPath)
 
       cli.action.start('Uploading app')
-
-      const appId = this.getAppID(appPath)
+      const appConfig = getAllConfigs(appPath) || {}
       const manifest = getManifestFile(appPath)
       const pkgPath = await createAppPkg(appPath)
       const { id: upload_id } = await uploadAppPkg(pkgPath)
@@ -77,7 +79,7 @@ export default class Update extends Command {
 
       cli.action.stop('Uploaded')
       try {
-        await this.installApp(appId, upload_id, appPath, manifest)
+        await this.installApp(appConfig, upload_id, appPath, manifest)
       } catch (error) {
         this.error(chalk.red(error))
       }
