+++ packages/zcli-apps/src/constants.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_APPS_CONFIG_FILE = 'zcli.apps.config.json'
