+++ packages/zcli-apps/src/lib/buildAppJSON.ts
@@ -18,7 +18,7 @@ import * as fs from 'fs'
 import { uuidV4 } from '../utils/uuid'
 import { getManifestFile } from '../utils/manifest'
 import { getAllConfigs } from '../utils/appConfig'
-import { promptAndGetSettings } from '../utils/createApp'
+import { getAppSettings } from '../utils/getAppSettings'
 import { validateAppPath } from './appPath'
 
 // The SVGs here are in the top bar or nav bar locations. Chat don’t have these locations thus not here.
@@ -85,29 +85,13 @@ export const getAppPayloadFromManifest = (appManifest: Manifest, port: number, a
   }
 }
 
-export const getAppSettings = async (manifest: Manifest, configParams: ConfigParameters) => {
-  if (!manifest.parameters) return {}
-  const configContainsParam = (paramName: string) => Object.keys(configParams).includes(paramName)
-
-  const paramsNotInConfig = manifest.parameters.filter(param => !configContainsParam(param.name))
-  const configSettings = manifest.parameters.reduce((result: Dictionary<string>, param) => {
-    if (configContainsParam(param.name)) {
-      result[param.name] = configParams[param.name] as string
-    }
-    return result
-  }, {})
-
-  const promptSettings = paramsNotInConfig ? await promptAndGetSettings(paramsNotInConfig, manifest.name, false) : {}
-  return { ...configSettings, ...promptSettings }
-}
-
-export const buildAppJSON = async (appPaths: string[], port: number, configFileName: string): Promise<AppJSONPayload> => {
+export const buildAppJSON = async (appPaths: string[], port: number): Promise<AppJSONPayload> => {
   const appJSON: AppJSON = { apps: [], installations: [] }
 
   for (const appPath of appPaths) {
     validateAppPath(appPath)
     const manifest = getManifestFile(appPath)
-    const zcliConfigFile = getAllConfigs(appPath, configFileName) || {}
+    const zcliConfigFile = getAllConfigs(appPath) || {}
 
     const appId = zcliConfigFile.app_id?.toString() || uuidV4()
     const configParams = zcliConfigFile.parameters || {} // if there are no parameters in the config, just attach an empty object
