+++ packages/zcli-apps/src/utils/appConfig.ts
@@ -4,8 +4,9 @@ import * as path from 'path'
 import * as fs from 'fs-extra'
 import * as chalk from 'chalk'
 import { CLIError } from '@oclif/errors'
+import { DEFAULT_APPS_CONFIG_FILE } from '../constants'
 
-export const getAllConfigs = (appPath: string, configFileName: string): ZcliConfigFileContent | undefined => {
+export const getAllConfigs = (appPath: string, configFileName: string = DEFAULT_APPS_CONFIG_FILE): ZcliConfigFileContent | undefined => {
   const configFilePath = path.join(appPath, configFileName)
 
   if (fs.existsSync(configFilePath)) {
